app/requirements.txt
arm-mango==1.3.1
attrdict==2.0.1
boto3==1.26.114
botocore==1.29.114
click==8.1.3
Flask==2.1.3
flask-lambda==0.0.4
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.2
pandas==2.0.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.0.3
zipp==3.15.0